% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model.r
\name{append_scorecard_time}
\alias{append_scorecard_time}
\title{append scorecard and time column at first}
\usage{
append_scorecard_time(df, run_date, scorecard)
}
\description{
append scorecard and time column at first
}
