% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofitr.r
\name{autofitr}
\alias{autofitr}
\title{autofitr Main Interface}
\usage{
autofitr(scorecard, server = "apprwidb006", db = "MAWORK_DSOL",
  lookup_table = "DSOL_Monitoring.dbo.monitoring_autofitr_lookup")
}
\arguments{
\item{scorecard}{scorecard name, should match to the [scorecard] column in lookup table}

\item{server}{sql server name for the lookup table and output tables, default as \code{apprwidb006}}

\item{db}{database for the lookup table and output tables, default as \code{MAWORK_DSOL}}

\item{lookup_table}{table name for the lookup table and output tables, default as \code{[CORP\\Mufeng.Zou].autofitr_lookup}}
}
\value{
List of metrics with length 4, where contents are identical to the output sql tables.
}
\description{
Automatically fit GBM models while applying credit scoring methodology and restrictions.
}
\details{
\code{autofitr} uses characteristics data from table specified by parameters,
applies pre-processing rules similarly to standard scorecard development (bucketing with enough bads, monotonic trends),
fits 3 GBM models with different variables/interaction-depths (devvar 1 depth, dev+new var 1 depth, dev+new 3 depths),
then write 4 metrics tables \code{myautofitr_value}, \code{myautofitr_gini}, \code{myautofitr_ca}, \code{myautofitr_corr} into server \code{server}, database \code{db}.

The output sql tables / list of metrics are:

\code{myautofitr_value}: fitted value from all 3 GBM models

\code{myautofitr_gini}: fitted Gini values from all 3 GBM models

\code{myautofitr_ca}: CAs for "devvar" model and "dev+new var" model.
(CA not available for "deeper" model, which has depths of 3)

\code{myautofitr_corr}: Correlation matrix from the "dev+new var" model,
including new variables newly joined the model, and all dev variables.
}
\examples{
\dontrun{

#to make results reproducible
set.seed(25252)

#using default lookup in apprwidb006 (e.g. for scorecard monitoring purposes)
autofitr("COMMPRS")

#if you want to use your own lookup, or on other server
#(to see how to create a lookup table, have a look at the default lookup table in apprwidb006)
autofitr("myscorecard", "myserver", "mydb", "mylookup")
}

}
\author{
Mufeng Zou, \email{mufeng.zou@equifax.com}
}
