% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.r
\name{bucket_func.factor}
\alias{bucket_func.factor}
\title{ca buckets for 1 factor vector}
\usage{
\method{bucket_func}{factor}(x, gbf, minbads = 20, maxbuckets = 20)
}
\description{
ca buckets for 1 factor vector
}
