% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.r
\name{bucket_func.numeric}
\alias{bucket_func.numeric}
\title{ca buckets for 1 numeric vector}
\usage{
\method{bucket_func}{numeric}(x, gbf, minbads = 20, maxbuckets = 20)
}
\description{
ca buckets for 1 numeric vector
}
