% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extract_devvar}
\alias{extract_devvar}
\title{extract devvar string from
long string for points calculations
long string for all chars available}
\usage{
extract_devvar(points_str, char_str)
}
\description{
extract devvar string from
long string for points calculations
long string for all chars available
}
\examples{
\dontrun{
#this can be a huge block of codes from points table creation
points_str <- "case  when defaults_y_12m>=1 then -1.3344   else 0.0	end as thvs11cs_defaults_y_12m"

#this is list of columns from char table
char_str <- "insolvencies	th_ind_dischargeFlag	th_ind_bankruptcyStatus_type defaults_y_12m"

extract_devvar(points_str, char_str)
}

}
