% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{gini}
\alias{gini}
\title{calculate gini from sql server table
return gini value, or data frame if "by" field is provided}
\usage{
gini(odbc, tbl, score, gbf, by, filter = "")
}
\description{
calculate gini from sql server table
return gini value, or data frame if "by" field is provided
}
\examples{
\dontrun{
gini('mawork_stg','myout','c090_master_scale','x_gbf_c090')

gini('mawork_stg','myout','c360_master_scale','x_gbf_c360')

gini(odbc='mawork_stg',tbl='nztemp_temp',score='fitted_value',gbf='gbf_temp',by='segment')
}
}
