% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{parse_args}
\alias{parse_args}
\title{0. parse arguments from cmd line, or from a vector of length 3
args <- c('apprwidb006', 'MAWORK_DSOL', 'monitoring_rebuild_testin2')}
\usage{
parse_args(args)
}
\description{
0. parse arguments from cmd line, or from a vector of length 3
args <- c('apprwidb006', 'MAWORK_DSOL', 'monitoring_rebuild_testin2')
}
