% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.r
\name{preprocess_vars}
\alias{preprocess_vars}
\title{2. preprocess dataframe}
\usage{
preprocess_vars(df, maxlevel = 100)
}
\description{
Assuming input dataframe has gbf column with value 'G'/'B',
this function will:
1. change 'G'/'B' to 1/0, while 1/0 stays the same
2. remove all missing and constant columns
3. convert numeric columns with only 1 valid value and NA to a factor (required for GBM)
}
