% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{read_from_sql}
\alias{read_from_sql}
\title{1. read inputs from sqlserver
assume all relevant tables are in same server
returns a list of dataframe with attr(,"segment") of score segments
and each dataframe has columns:
user_ref, gbf, devvar, newvar}
\usage{
read_from_sql(scorecard, lookup_server, lookup_db, lookup_table)
}
\description{
1. read inputs from sqlserver
assume all relevant tables are in same server
returns a list of dataframe with attr(,"segment") of score segments
and each dataframe has columns:
user_ref, gbf, devvar, newvar
}
