% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.r
\name{write_to}
\alias{write_to}
\title{5. output 3 tables:
myautofitr_value: fitted value from different models
myautofitr_gini: devvar gini and dev+newvar gini
myautofitr_ca: CA and variable importance
myautofitr_corr: correlation matrix newvar to devvar}
\usage{
write_to(sqlserver_name, db_name, metrics)
}
\description{
5. output 3 tables:
myautofitr_value: fitted value from different models
myautofitr_gini: devvar gini and dev+newvar gini
myautofitr_ca: CA and variable importance
myautofitr_corr: correlation matrix newvar to devvar
}
